requirements.txt
django
django-split-settings
djangorestframework
djangorestframework_simplejwt
pillow
pysaml2